+++ dbus-serialbattery/bms/lltjbd.py
@@ -239,8 +239,8 @@ class LltJbd(Battery):
         self.type = self.BATTERYTYPE
         self.address = address
         self._product_name: str = ""
-        self.has_settings = False
-        self.reset_soc = 100
+        self.has_settings = True
+        self.soc_reset_to = 100
         self.soc_to_set = None
         self.factory_mode = False
         self.writable = False
@@ -249,10 +249,11 @@ class LltJbd(Battery):
         self.trigger_disable_balancer = None
         self.cycle_capacity = None
         # list of available callbacks, in order to display the buttons in the GUI
-        self.available_callbacks = [
-            "force_charging_off_callback",
-            "force_discharging_off_callback",
-            "turn_balancing_off_callback",
+        self.callbacks_available = [
+            "callback_charging_force_off",
+            "callback_discharging_force_off",
+            "callback_balancing_turn_off",
+            "callback_soc_reset_to",
         ]
         self.history.exclude_values_to_calculate = ["charge_cycles"]
 
@@ -325,14 +326,14 @@ class LltJbd(Battery):
 
         return True
 
-    def reset_soc_callback(self, path, value):
+    def callback_soc_reset_to(self, path, value):
         if value is None:
             return False
 
         if value < 0 or value > 100:
             return False
 
-        self.reset_soc = value
+        self.soc_reset_to = value
         self.soc_to_set = value
 
         return True
@@ -350,7 +351,7 @@ class LltJbd(Battery):
             pack_voltage = struct.pack(">H", int(self.voltage * 10))
             self.read_serial_data_llt(writeCmd(REG_CAP_100, pack_voltage))
 
-    def force_charging_off_callback(self, path, value):
+    def callback_charging_force_off(self, path, value):
         if value is None:
             return False
 
@@ -364,7 +365,7 @@ class LltJbd(Battery):
 
         return False
 
-    def force_discharging_off_callback(self, path, value):
+    def callback_discharging_force_off(self, path, value):
         if value is None:
             return False
 
@@ -408,7 +409,7 @@ class LltJbd(Battery):
             logger.error("write force disable charge/discharge failed")
             return False
 
-    def turn_balancing_off_callback(self, path, value):
+    def callback_balancing_turn_off(self, path, value):
         if value is None:
             return False